requirements.txt
schedule
mysql-connector-python
convertdate

python-dotenv
pyTelegramBotAPI
sqlalchemy
pymysql 
coloredlogs